# backend/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Parbati Enterprises Expense Tracker API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /vouchers/payment:
    post:
      operationId: createPaymentVoucher
      tags: [vouchers]
      summary: Create a payment voucher (expense_entry+)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVoucherRequest"
      responses:
        "201":
          description: Payment voucher created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VoucherDetailResponse"
  /vouchers/receive:
    post:
      operationId: createReceiveVoucher
      tags: [vouchers]
      summary: Create a receive voucher (expense_entry+)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVoucherRequest"
      responses:
        "201":
          description: Receive voucher created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VoucherDetailResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    VoucherDetail:
      type: object
      properties:
        expenseId:
          type: integer
        voucherType:
          type: string
          enum: [payment, receive]
        projectId:
          type: integer
        projectName:
          type: string
        vendorId:
          type: integer
        vendorName:
          type: string
        createdBy:
          type: integer
        approvedBy:
          type: integer
          nullable: true
        approvedAt:
          type: string
          format: date-time
          nullable: true
        finalizedBy:
          type: integer
          nullable: true
        finalizedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        versionId:
          type: integer
        versionNo:
          type: integer
        voucherNumber:
          type: string
        expenseDate:
          type: string
          format: date
        particularId:
          type: integer
        particularName:
          type: string
        description:
          type: string
          nullable: true
        transactionDetails:
          type: string
          nullable: true
        uomId:
          type: integer
        uomName:
          type: string
        quantity:
          type: string
        pricePerUnit:
          type: string
        amount:
          type: string
        invoiceNumber:
          type: string
          nullable: true
        paymentStatusId:
          type: integer
        paymentStatusName:
          type: string
      required:
        - expenseId
        - voucherType
        - projectId
        - projectName
        - vendorId
        - vendorName
        - createdBy
        - createdAt
        - updatedAt
        - versionId
        - versionNo
        - voucherNumber
        - expenseDate
        - particularId
        - particularName
        - uomId
        - uomName
        - quantity
        - pricePerUnit
        - amount
        - paymentStatusId
        - paymentStatusName
    VoucherDetailResponse:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/VoucherDetail"
      required: [data]
...
    CreateVoucherRequest:
      type: object
      properties:
        projectId:
          type: integer
        vendorId:
          type: integer
        expenseDate:
          type: string
          format: date
        particularId:
          type: integer
        description:
          type: string
        transactionDetails:
          type: string
        uomId:
          type: integer
        quantity:
          type: number
        pricePerUnit:
          type: number
        invoiceNumber:
          type: string
        paymentStatusId:
          type: integer
      required:
        - projectId
        - vendorId
        - expenseDate
        - particularId
        - uomId
        - quantity
        - pricePerUnit
        - paymentStatusId
...